+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## release-next
 
 - Capture the fact a project may not exist when trying to view it.
+- Render Workspace on the `project view [id]` command.
+- Render the application link to the project view.
 
 ## 0.0.42